Add files via upload
If used by others, you will have to change the folder paths for the files that will be created within the code. It is designed to search for a file within a path, and if it is not there, it will create a file. The code has a "users" directory. I suggest adding your own users directory to the same directory as the main.py

quickly search for file path >> /home/covid2020/Desktop/Users/