Small pull request1
Hi. I'm Agreeable_Big_9620 from reddit. I noticed that when entering spending information into your data, entering nothing results in a value error. If you accidently enter into the spending information section, the only way to get out is to restart the program. I decided to fix that, in the unlikely event that it is an accident.